add knowledge_search to be always included (#28)
knowledge_search tool is needed to query knowledge base, it gets excluded by agent filtering.
This still need a work todo to customize the tools that we should always include and also to
customize the filter prompt by env variable.
Also replace the logging debug to info, to have more logs to analyse when deployed.

Signed-off-by: Djebran Lezzoum <ldjebran@gmail.com>